support qwenvl2
